% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fh_distint_teams.R
\docType{data}
\name{fh_distinct_teams}
\alias{fh_distinct_teams}
\title{Field Hockey Distinct Team Names}
\format{
A character vector with 300 elements:
}
\usage{
fh_distinct_teams
}
\description{
A dataset containing distinct team names for women's field hockey teams
in the NCAAA
}
\keyword{datasets}
