% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fh_master_lu_table.R
\docType{data}
\name{fh_master_lu_table}
\alias{fh_master_lu_table}
\title{Master NCAA Field Hockey Team Lookup}
\format{
A data frame
}
\usage{
fh_master_lu_table
}
\description{
A data frame containing a master list of women's field hockey teams, team IDs,
conferences, and conference IDs
}
\keyword{datasets}
