% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fh_year_lu_table.R
\docType{data}
\name{fh_year_lu_table}
\alias{fh_year_lu_table}
\title{NCAA Women's Field Hockey Season ID Lookup Table}
\format{
A data frame
}
\usage{
fh_year_lu_table
}
\description{
A data frame containing a master list of women's field hockey season IDs
}
\keyword{datasets}
