% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_ncaa_wfh_team_schedules.R
\name{get_ncaa_wfh_team_schedules}
\alias{get_ncaa_wfh_team_schedules}
\title{Acquire NCAA Women's Field Hockey Schedules and Results}
\usage{
get_ncaa_wfh_team_schedules(team_id, year, get_boxscore_urls = FALSE)
}
\arguments{
\item{team_id}{The unique team id from the NCAA}

\item{year}{The season for which you want data.
If you want the 2017-18 season, you would use 2018.}
}
\value{
A data frame
}
\description{
Acquire NCAA Women's Field Hockey Schedules and Results
}
\examples{
get_ncaa_wfh_team_schedules(712, 2018)
}
