% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/school_lu.R
\name{school_lu}
\alias{school_lu}
\title{Look up Team Information}
\usage{
school_lu(school_name, school_division, school_conference)
}
\arguments{
\item{school_name}{Full or partial name of the school (string)}

\item{school_division}{Division of the school (numeric)}

\item{school_conference}{Full or partial name of the conference (string)}
}
\value{
A data frame
}
\description{
Look up Team Information
}
\examples{
\dontrun{school_lu(school_name = 'California', school_division = 1, school_conference = 'PAC')}
}
